requirements.txt
cheroot==8.3.0
colorama==0.4.3
jaraco.functools==3.0.0
Markdown==3.2.1
more-itertools==8.2.0
Pillow==7.0.0
qrcode==6.1
six==1.14.0
web.py==0.40
